Move project helpers to new file
